// src/types/bindings/OutstandingReport.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CurrencyAmount } from "./CurrencyAmount";
import type { OutstandingInvoice } from "./OutstandingInvoice";

export type OutstandingReport = { 
/**
 * Sorted by days_overdue desc, then by due_date asc.
 */
invoices: Array<OutstandingInvoice>, total_outstanding: Array<CurrencyAmount>, };